moved player module to player.js
