+++ player.js
@@ -1,3 +1,6 @@
+var Scrabble = require('./Scrabble.js')
+
+
 var playedWordsArray = []
 
 var Player = function(name = "Player") {
@@ -9,7 +12,7 @@ Player.prototype.plays = function(playedWords){
   playedWordsArray = playedWords;
   }
   return playedWordsArray;
-}
+};
 
 Player.prototype.play = function(word){
   playedWordsArray.push(word);
@@ -18,29 +21,50 @@ Player.prototype.play = function(word){
   } else {
   return playedWordsArray;
   }
-}
+};
 
 Player.prototype.totalScore = function(){
-
-}
+  var playedWords = this.plays();
+  var total = 0
+  for (var n = 0, len = playedWords.length; n < len; n++) {
+    total += Scrabble.score(playedWords[n]);
+  }
+  return total;
+};
 
 Player.prototype.hasWon = function(){
-
-}
+  var total = this.totalScore();
+  if (total >= 100){
+    return true
+  } else {
+    return false;
+  }
+};
 
 Player.prototype.highestScoringWord = function(){
-
+ var playedWords = this.plays();
+ var topWord = Scrabble.highestScoreFrom(playedWords);
+ return topWord;
 }
 
 Player.prototype.highestWordScore = function(){
-
+ topScore = Scrabble.score(this.highestScoringWord());
+ return topScore;
 }
 
+
+
+
 var playerOne = new Player("Tom");
-var array = ["happy", "sad", "angry"];
+var array = ["dog", "fkkaaaa", "gdog", "aaaajhd", "aaaaqgg", "cat", "dogdo", "catca"];
 console.log(playerOne.name);
-console.log(playerOne.plays(array));
+playerOne.plays(array);
 playerOne.play("cat");
-console.log(playerOne.plays(array));
 playerOne.play("dog");
-console.log(playerOne.plays(array));
+// console.log(playerOne.plays(array));
+
+
+console.log(playerOne.totalScore());
+console.log("has won??  " + playerOne.hasWon());
+console.log(playerOne.highestScoringWord());
+console.log(playerOne.highestWordScore());
